uploaded files tab base ui updated. moving on to results display ui
