--- src/jsMain/kotlin/css/component/fileupload/FileUploadButtonBarStyle.kt
@@ -1,9 +1,0 @@
-package css.component.fileupload
-
-import kotlinx.css.*
-import styled.StyleSheet
-
-object FileUploadButtonBarStyle : StyleSheet("FileUploadButtonBar", isStatic = true) {
-
-
-}
