+++ src/jsMain/kotlin/css/tabs/uploadtab/filelist/FileStatusSpinnerStyle.kt
@@ -1,77 +1,20 @@
-package css.component.fileupload.filelist
+package css.tabs.uploadtab.filelist
 
+import css.component.FileItemStyle
 import css.const.*
 import kotlinx.css.*
 import kotlinx.css.properties.*
 import styled.StyleSheet
 import styled.animation
 
-object FileStatusIndicatorStyle : StyleSheet("FileStatusIndicator", isStatic = true) {
-
-    val indicator by css {
-        display = Display.inlineBlock
-        alignItems = Align.center
-        boxSizing = BoxSizing.borderBox
-        justifyContent = JustifyContent.center
-        height = 32.px
-        width = 32.px
-        borderRadius = 50.pct
-    }
-
-    val animatedIndicator by css {
-        +indicator
-        scaleIntro()
-    }
-
-    val indicatorNoStatus by css {
-        +indicator
-        boxSizing = BoxSizing.borderBox
-        border(
-            width = 2.px,
-            style = BorderStyle.solid,
-            color = BORDER_GRAY)
-    }
-
-    val indicatorFatal by css {
-        +animatedIndicator
-        backgroundColor = FATAL_PINK
-    }
-
-    val indicatorError by css {
-        +animatedIndicator
-        backgroundColor = ERROR_ORANGE
-    }
-
-    val indicatorWarning by css {
-        +animatedIndicator
-        backgroundColor = WARNING_YELLOW
-    }
-
-    val indicatorInformation by css {
-        +animatedIndicator
-        backgroundColor = SUCCESS_GREEN
-    }
-
-    val indicatorGood by css {
-        +animatedIndicator
-        backgroundColor = SUCCESS_GREEN
-    }
-
-    val imageContainer by css {
-        width = 100.pct
-        height = 100.pct
-        display = Display.flex
-        alignItems = Align.center
-        justifyContent = JustifyContent.center
-    }
-
-    val indicatorImage by css {
-        width = 80.pct
-        height = 80.pct
-    }
+object FileStatusSpinnerStyle : StyleSheet("FileStatusSpinnerStyle", isStatic = true) {
 
     val loadingIndicator by css {
-        +indicator
+        display = Display.inlineBlock
+        boxSizing = BoxSizing.borderBox
+        width = 100.pct
+        height = 100.pct
+        borderRadius = 50.pct
         border(width = 2.px, style = BorderStyle.solid, color = HIGHLIGHT_GRAY)
         borderTop(width = 2.px, style = BorderStyle.solid, color = BORDER_GRAY)
         spinner()
@@ -98,7 +41,7 @@ object FileStatusIndicatorStyle : StyleSheet("FileStatusIndicator", isStatic = t
 
     fun CSSBuilder.scaleOutro() {
         animation(
-            duration = (0.25).s,
+            duration = (0.5).s,
             timing = Timing.materialAcceleration,
         ) {
             this.run {
@@ -131,5 +74,20 @@ object FileStatusIndicatorStyle : StyleSheet("FileStatusIndicator", isStatic = t
                 ))
             }
         }
+    }
+
+
+
+    //TODO move this to another file
+
+    val loadingIconLight by css {
+        height = 2.rem
+        width = 2.rem
+        borderRadius = 50.pct
+        display = Display.inlineBlock
+        marginBottom = 1.rem
+        border(width = FileItemStyle.INDICATOR_BORDER_WIDTH, style = BorderStyle.solid, color = ALMOST_RED, borderRadius = 50.pct)
+        borderTop(width = FileItemStyle.INDICATOR_BORDER_WIDTH, style = BorderStyle.solid, color = REALLY_RED)
+        spinner()
     }
 }
